feature: update parameters's type

+++ examples/server.py
@@ -508,8 +508,8 @@ def update_mpr_type(
 @app.get('/panmpr')
 def pan_mpr(
         plane_type: int,
-        x: int,
-        y:int
+        x: float,
+        y: float
 ):
     hm.PanCrossHair(x, y, mk.PlaneType(plane_type))
